src/main/java/com/thoughtworks/capability/service/ShoppingCartService.java
package com.thoughtworks.capability.service;

import com.thoughtworks.capability.repository.ProductRepository;
import com.thoughtworks.capability.web.dto.ShoppingCartResponse;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.math.BigDecimal;
import java.util.List;

@Service
@RequiredArgsConstructor
public class ShoppingCartService {
    private final ProductRepository productRepository;

    public ShoppingCartResponse findShoppingCart() {
        List products = productRepository.findAll();
        return new ShoppingCartResponse(products, BigDecimal.ZERO);
    }
}
